test of backend and frontend for VAG-8 implemented(bring ingredients)
